Moved logout button inside sidebar
