+++ src/frontend/src/components/layout/Sidebar.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 import { slide as Menu } from "react-burger-menu";
+import Logout from "./Logout.js";
 
 // HAMBURGER NAV ITEMS (HOME)
 // Used this object structure to create the sidebar menu items:
@@ -15,25 +16,30 @@ import { slide as Menu } from "react-burger-menu";
 */
 function Sidebar({ items }) {
   return (
-    <Menu>
-      {items.map(({ title, contents }) => (
-        <div className="hamburger-nav">
-          {title}
-          {contents.map(({ destination, label }) => (
-            <li>
-              <NavLink
-                key={label}
-                className="menu-item"
-                to={destination}
-                style={{ cursor: "pointer" }}
-              >
-                {label}
-              </NavLink>
-            </li>
-          ))}{" "}
-        </div>
-      ))}
-    </Menu>
+    <div>
+      <Menu>
+        {items.map(({ title, contents }) => (
+          <div className="hamburger-nav">
+            {title}
+            {contents.map(({ destination, label }) => (
+              <li>
+                <NavLink
+                  key={label}
+                  className="menu-item"
+                  to={destination}
+                  style={{
+                    cursor: "pointer",
+                  }}
+                >
+                  {label}
+                </NavLink>
+              </li>
+            ))}{" "}
+          </div>
+        ))}
+        <Logout />
+      </Menu>
+    </div>
 
     // <Menu>
     //   {items.map(({ label, destination }) => (
